Plip submodule, lig_interactions function w/ cli
Folded the plip module into ligule so there is no confusion with the
plip module for download and so interpackage dependencies are not an
issue. Refactored module names while I was at it.

Made a new ligule command for profiling ligand interactions. Currently
only supports hydrogen bond profiling. Supports one type of output which
is to make a dataframe where columns are the ligand atoms (non-hydrogen)
and the rows are different input complex structures. 1s are placed where
the interaction type is found for a ligand atom in a structure and 0 for
no interaction.

Built as a framework to take in different interaction types and output
types at once and output them all, so it can be extended pretty easy,
for instance reporting the distances of pairs with an interaction.