Visualization, weighted interaction frequencies.
Added function to generate frequencies of interactions for specific
atoms of a ligand. Then you can sum them or do a weighted sum and output
as a small csv. I then added functions to apply these values to the
b-factor column of a pdb so that it can be visualized.